Allow setting a typemap

+++ test/migrations/c/001_init.ts
@@ -3,7 +3,7 @@ import type { Sql } from "../../../src";
 export async function up(sql: Sql) {
   await sql`
     create table test_migrations.users (
-      id serial primary key,
+      id bigserial primary key,
       first_name text not null,
       last_name text not null
     );
@@ -11,9 +11,9 @@ export async function up(sql: Sql) {
 
   await sql`
     create table test_migrations.notes (
-      id serial primary key,
+      id bigserial primary key,
       body text not null,
-      author_id int not null references test_migrations.users(id),
+      author_id bigint not null references test_migrations.users(id),
       created_at timestamp not null default now()
     );
   `.exec();
